Bellman–Ford Algorithm
shortest paths from src to all vertices in the given graph. The graph may contain negative weight edges. 
We have discussed Dijkstra’s algorithm for this problem. 
Dijkstra’s algorithm is a Greedy algorithm and the time complexity is O((V+E)LogV) (with the use of the Fibonacci heap).